backend/requirements.txt
fastapi
uvicorn[standard]
pyserial
strawberry-graphql